rearrange all of Query.ts
