+++ src/resolvers/Query/index.ts
@@ -1,0 +1,12 @@
+// @TODO make all these functions conform to the same
+// graphql resolver signature
+//
+// @see https://gist.github.com/sampsyo/7f1fa4f2ebc10088a7d
+export { getAllAttempts } from "./getAllAttempts"
+export { getAllComments } from "./getAllComments"
+export { getAllReactions } from "./getAllReactions"
+export { getAllUsers } from "./getAllUsers"
+export { getAttemptsByUserId } from "./getAttemptsByUserId"
+export { getCommentsByUrl } from "./getCommentsByUrl"
+export { getFirstUser } from "./getFirstUser"
+export { getUserById } from "./getUserById"
